SBERDOMA-20 move search function to module, move Ticket.uistate.js to utils/clientScheema
