+++ apps/condo/containers/TicketForm/CreateTicketForm.tsx
@@ -3,8 +3,8 @@ import { useOrganization } from '@core/next/organization'
 import { useRouter } from 'next/router'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
-import { useCreate } from '../../schema/Ticket.uistate'
-import { BaseTicketForm } from './BaseTicketForm'
+import { useCreate } from '../../utils/clientSchema/Ticket'
+import { BaseTicketForm } from '../../components/BaseTicketForm'
 
 const OPEN_STATUS = '6ef3abc4-022f-481b-90fb-8430345ebfc2'
