+++ apps/condo/pages/ticket/[id]/update.tsx
@@ -4,7 +4,7 @@ import Head from 'next/head'
 import { useIntl } from '@core/next/intl'
 import { PageContent, PageHeader, PageWrapper } from '../../../containers/BaseLayout'
 import { OrganizationRequired } from '../../../containers/OrganizationRequired'
-import { TicketForm } from '../../../components/TicketForm'
+import { TicketForm } from '../../../containers/TicketForm'
 
 const TicketUpdatePage = () => {
     const intl = useIntl()
