+++ apps/condo/utils/clientSchema/Ticket.ts
@@ -1,7 +1,7 @@
 import { genReactHooks } from '@core/keystone/gen.gql.react.utils'
 
-import { Ticket } from './Ticket.gql'
-import { getClientSideSenderInfo } from '../utils/userid.utils'
+import { Ticket } from '../../schema/Ticket.gql'
+import { getClientSideSenderInfo } from '../userid.utils'
 
 function convertGQLItemToUIState (item) {
     if (item.dv !== 1) throw new Error('unsupported item.dv')
